refactor: update test descriptions and restructure provider imports
