+++ renderer/__test__/App.test.tsx
@@ -7,7 +7,7 @@ import { renderWithProvider } from './utils';
 
 describe('App', () => {
   describe('Routing', async () => {
-    it('home without locale', async () => {
+    it('Home Page Render', async () => {
       renderWithProvider(<App routerType="memory" initialEntries={['/']} />);
       expect(await screen.findByTestId('home-page')).toBeInTheDocument();
     });
