+++ renderer/__test__/utils.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Providers from '@/Providers';
+import Providers from '@/providers/AppProviders';
 
 export function renderWithProvider(children: React.ReactNode) {
   return render(<Providers>{children}</Providers>);
